Move make and all overloads to `registration.py`
